requirements.txt
Adafruit-Blinka==8.47.0
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-pn532==2.4.1
adafruit-circuitpython-requests==4.1.6
adafruit-circuitpython-typing==1.11.0
Adafruit-PlatformDetect==3.74.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
cachetools==5.5.0
certifi==2020.6.20
chardet==4.0.0
colorzero==1.1
distro==1.5.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
gpiozero==1.6.2
gspread==6.1.2
httplib2==0.22.0
idna==2.10
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.2.2
picamera2==0.3.12
pidng==4.0.9
piexif==1.1.3
Pillow==8.1.2
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyftdi==0.55.4
pyparsing==3.1.4
pyserial==3.5
python-apt==2.2.1
python-prctl==1.7
pyusb==1.2.1
requests==2.25.1
requests-oauthlib==2.0.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.0
rsa==4.9
simplejpeg==1.6.4
six==1.16.0
spidev==3.5
ssh-import-id==5.10
sysv-ipc==1.1.0
toml==0.10.1
typing-extensions==4.12.2
urllib3==1.26.5
v4l2-python3==0.3.2
